Added FAT32 Filesystem

+++ drivers/disk.h
@@ -9,6 +9,6 @@
 
 int disk_init(void);
 
-int ATA_read_sectors(uint8_t drive, uint32_t start_sector, uint32_t count, void* buffer);
+int8_t ATA_read_sectors(uint8_t drive, uint32_t start_sector, uint32_t count, void* buffer);
